+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [unreleased]
 
+## [0.4.0]
+
 - Switched QR code scanning to scan full page and report orientation
 - Added second attempt for QR code scanning with higher default resolution, `try_rotate=True`, and `try_downscale=True`
 - Removed dependency on `scipy`